feat: Add await keyword to addBattle function
This commit adds the `await` keyword to the `addBattle` function in the `battle.ts` file. By using `await`, we ensure that the `db.collection('battles').add()` operation is completed before moving on to the next line of code. This change ensures that the battle is successfully added to the Firestore collection before any subsequent actions are taken.